Merged with calenar
I merged in the calendar_vim_reminder program I had also written. As a
result the daily digest now shows calendar events. The idea for this is
that I want to have a module type system so I can just put in a module
for every new thing I want to add to my daily digest emails.