Created abstract class (ish) so code reorginization TODO was modified to reflect new goals
Digest was modified to reflect changes in calendarvim

source_interface is my attempt to make an interface for all of the
modules that will be going into the digest. At the moment I still
haven't completely figured out how to modularize each source, but this
is my first attempt.

calendarvim is the first change with the new interface, and most of it
was updated accordingly. That includes bringing events and calendar
classes under CalendarVim object and added the get_digest methods as
well as trying to make most of the methods private outside of
get_digest.